# backend/spring-infra/src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: Flexstore API
  description: REST contracts for Flexstore (hexagonal architecture online store sample project)
  contact:
    name: Emmanuel Olivier
    url: https://github.com/eolivier
  version: v0.1.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/cart/save:
    post:
      tags:
      - Cart
      summary: Add an item to the cart
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonItem"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    JsonItem:
      type: object
      properties:
        itemId:
          type: string
        productId:
          type: string
        productName:
          type: string
        productDescription:
          type: string
        productCategory:
          type: string
        productPrice:
          type: number
        productCurrency:
          type: string
        productQuantity:
          type: integer
          format: int32
...